Allow cargo builds against other branches of rust
Cargo itself doesn't yet have the concept of stable/beta/nightly as it is still
pre 1.0.0. However we can build cargo with beta and stable versions of Rust to
ensure compatability between the the compiler and the package manager. This
update is the first in that step and supports stable/beta/nightly (and
individual tags theoretically)

Also there are a few compatability changes and container changes with this
release. The cargo and rust builds are now stored in
/opt/rust_{stable,beta,nightly}/{cargo,rust} for design purposes. This is also
the first step in deploying multirust to the build system and using that to
manage the paths for rust and cargo instead of handling pathing manually.

This is an unstable commit, the builds work, but not everything has been
fully debugged to call this stable. To update your existing container to
support this build, run mkdir -p /opt/rust_{stable,beta,nightly}/{cargo,rust}
from within a chroot or systemd-nspawn within the container and
rm -r /build/nightly to remove the old nightly directory.