Added a proxy script for dropbox
Added a proxy script to cache the uploads and the downloads through the
dropbox script. This will be usefull in reducing the time we need to wait
for downloads of files we've recently built.

This script should have some sort of max size parameter, so that the we
can cull the oldest files when it gets larger than our desired cache size.