Creating scripts to build a raspbian container
Added a CONTAINER_TAG to the environment for the usage of differentiating
between the builds between containers.

Tweaked the raspbian creation and setup, following the basic ideas presented in
http://elinux.org/Raspbian

Continuing work on raspbian container setup. I think I will only build on the
raspbian container only occasionally as I don't have a RPi to test on.