Fixed the newer build process
Updated the the caching script to avoid the old seq, and use the newer
Bash for loop expression. Updated the build scripts to reference the
correct directories. Re-added the call to "git checkout ." to resolve
the issue with not being able to apply the static-ssl patch because
we had already applied it in a previous run. There's probably a
better fix for this, but until then, this will work nicely.