Performance work
Using a standlaone vanilla LLVM instead of the Rust built one. Hope it improves the build time
Got rid of a bunch of scripts for the soft float test against a really old glibc. It wasn't working.
Reconfigured the debian setup to compile a new CMake and LLVM+Clang for the container.